# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WeWorkout API",
        "description": "Documentación de la API del backend de la aplicación WeWorkout. La mayoría de llamadas utilizan autenticación mediante JWT. Se incluyen ejemplos de request y de response.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Inicia sesión",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "ironman@avengers.com",
                                    "password": "PepperPotts<3"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK. Devuelve el token y datos del usuario",
                        "content": {
                            "application/json": {
                                "example": {
                                    "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC8xMjcuMC4wLjE6ODAwMFwvYXBpXC9hdXRoXC9sb2dpbiIsImlhdCI6MTYyMjU0NzU5MywiZXhwIjoxNjIyNTUxMTkzLCJuYmYiOjE2MjI1NDc1OTMsImp0aSI6ImZsTDhVeVptc0pncFlIZDMiLCJzdWIiOjEsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.qHb0koGvpInuyMlVWyazc7DxgFALfX0uFLzpi1H4KDg",
                                    "token_type": "bearer",
                                    "expires_in": 3600,
                                    "user": {
                                        "id": 1,
                                        "name": "Carlos",
                                        "email": "ironman@avengers.com",
                                        "email_verified_at": "2021-05-31T21:11:08.000000Z",
                                        "birthdate": "1996-05-17T00:00:00.000000Z",
                                        "genre": "Hombre",
                                        "description": "Desc ejemplo",
                                        "created_at": "2021-05-31T21:11:08.000000Z",
                                        "updated_at": "2021-06-01T08:19:45.000000Z",
                                        "favourite_routines": [
                                            5,
                                            31
                                        ],
                                        "followers": [
                                            6
                                        ],
                                        "followees": [],
                                        "routines": [
                                            4,
                                            22,
                                            27,
                                            35,
                                            38,
                                            41
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    }
                }
            }
        },
        "/api/auth/me": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtiene información del usuario",
                "responses": {
                    "200": {
                        "description": "OK. Devuelve datos del usuario",
                        "content": {
                            "application/json": {
                                "example": {
                                    "user": {
                                        "id": 1,
                                        "name": "Carlos",
                                        "email": "ironman@avengers.com",
                                        "email_verified_at": "2021-05-31T21:11:08.000000Z",
                                        "birthdate": "1996-05-17T00:00:00.000000Z",
                                        "genre": "Hombre",
                                        "description": "Desc ejemplo",
                                        "created_at": "2021-05-31T21:11:08.000000Z",
                                        "updated_at": "2021-06-01T08:19:45.000000Z",
                                        "favourite_routines": [
                                            5,
                                            31
                                        ],
                                        "followers": [
                                            6
                                        ],
                                        "followees": [],
                                        "routines": [
                                            4,
                                            22,
                                            27,
                                            35,
                                            38,
                                            41
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cierra sesión",
                "responses": {
                    "200": {
                        "description": "OK. Cierra sesión",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Sesión cerrada con éxito"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}